Address PR review comments for master-effects fix
