+++ src/project/tests/CMakeLists.txt
@@ -10,6 +10,7 @@ set(MODULE_TEST_SRC
     ${CMAKE_CURRENT_LIST_DIR}/mocks/audacityprojectmock.h
     ${CMAKE_CURRENT_LIST_DIR}/mocks/projectconfigurationmock.h
     ${CMAKE_CURRENT_LIST_DIR}/mocks/trackeditprojectcreatormock.h
+    ${CMAKE_CURRENT_LIST_DIR}/mocks/dummyeffectinstancefactory.h
     ${CMAKE_CURRENT_LIST_DIR}/audacityproject_tests.cpp
     ${CMAKE_CURRENT_LIST_DIR}/realtimeeffectspersistence_tests.cpp
 )
